back/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
add-trailing-comma==2.5.1
    # via -r requirements.in
asgiref==3.7.2
    # via django
attrs==23.1.0
    # via jsonschema
backports-zoneinfo==0.2.1
    # via django
black==23.3.0
    # via -r requirements.in
click==8.1.3
    # via black
django==4.2.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==4.0.0
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
drf-spectacular==0.26.2
    # via -r requirements.in
factory-boy==3.2.1
    # via -r requirements.in
faker==18.11.1
    # via
    #   -r requirements.in
    #   factory-boy
flake8==6.0.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
importlib-metadata==6.6.0
    # via markdown
importlib-resources==5.12.0
    # via jsonschema
inflection==0.5.1
    # via drf-spectacular
isort==5.12.0
    # via -r requirements.in
jsonschema==4.17.3
    # via drf-spectacular
markdown==3.4.3
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via black
pathspec==0.11.1
    # via black
pkgutil-resolve-name==1.3.10
    # via jsonschema
platformdirs==3.8.0
    # via black
psycopg2-binary==2.9.6
    # via -r requirements.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyjwt==2.7.0
    # via djangorestframework-simplejwt
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via faker
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tokenize-rt==5.1.0
    # via add-trailing-comma
tomli==2.0.1
    # via black
typing-extensions==4.6.3
    # via
    #   asgiref
    #   black
uritemplate==4.1.1
    # via drf-spectacular
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
